requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
pydantic[email]==2.10.4
pydantic-settings==2.7.1
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
cryptography==44.0.0
pyotp==2.9.0
qrcode[pil]==8.0
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
httpx==0.28.1
aiosqlite==0.20.0
